requirements.txt
playwright>=1.41.0
gspread>=6.1.0
google-auth>=2.28.0
